Refactor permission handling: Replace LuckPermsUtil with new PermissionUtil for permission checks and enhance notification handling in MessageProcessor
